+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - #271
 - #221
 
+### Added
+- `Transform.Move(double x, double y, double z)`
+- `Transform.Rotate(double angle)`
+
 ## 0.6.2
 ### Added
 - `Material.Unlit`